독후감 삭제 기능 수정
