--- app/src/main/java/com/example/mobilesw/adapter/BoardDeleter.java
@@ -1,88 +1,0 @@
-package com.example.mobilesw.adapter;
-
-//Firestore의 db와 연동, 게시글 삭제를 담당
-
-import android.app.Activity;
-
-import androidx.annotation.NonNull;
-
-import com.example.mobilesw.info.PostInfo;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.storage.FirebaseStorage;
-import com.google.firebase.storage.StorageReference;
-
-import java.util.ArrayList;
-
-import static com.example.mobilesw.info.Util.isStorageUrl;
-import static com.example.mobilesw.info.Util.showToast;
-import static com.example.mobilesw.info.Util.storageUrlToName;
-
-public class BoardDeleter {
-    private Activity activity;
-    private OnPostListener onPostListener;
-    private int successCount;
-
-    public BoardDeleter(Activity activity) {
-        this.activity = activity;
-    }
-
-    public void setOnPostListener(OnPostListener onPostListener){
-        this.onPostListener = onPostListener;
-    }
-
-    public void storageDelete(final PostInfo postInfo){
-        FirebaseStorage storage = FirebaseStorage.getInstance();
-        StorageReference storageRef = storage.getReference();
-
-        final String id = postInfo.getId();
-        ArrayList<String> contentsList = postInfo.getContents();
-        for (int i = 0; i < contentsList.size(); i++) {
-            String contents = contentsList.get(i);
-            if (isStorageUrl(contents)) {
-                successCount++;
-                System.out.println("successCount");
-                StorageReference desertRef = storageRef.child("posts/" + id + "/" + storageUrlToName(contents));
-                System.out.println("urlName: "+storageUrlToName(contents));
-                desertRef.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-                        successCount--;
-                        storeDelete(id, postInfo);
-                    }
-                }).addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception exception) {
-                        showToast(activity, "Error");
-                    }
-                });
-            }
-        }
-        storeDelete(id, postInfo);
-    }
-
-
-    private void storeDelete(final String id, final PostInfo postInfo) {
-        FirebaseFirestore firebaseFirestore = FirebaseFirestore.getInstance();
-        if (successCount == 0) {
-            firebaseFirestore.collection("posts").document(id)
-                    .delete()
-                    .addOnSuccessListener(new OnSuccessListener<Void>() {
-                        @Override
-                        public void onSuccess(Void aVoid) {
-                            showToast(activity, "게시글을 삭제하였습니다.");
-                            onPostListener.onDelete(postInfo);
-                            //postsUpdate();
-                        }
-                    })
-                    .addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            showToast(activity, "게시글을 삭제하지 못하였습니다.");
-                        }
-                    });
-        }
-    }
-
-}
