--- app/src/main/java/com/example/mobilesw/adapter/OnPostListener.java
@@ -1,8 +1,0 @@
-package com.example.mobilesw.adapter;
-
-import com.example.mobilesw.info.PostInfo;
-
-public interface OnPostListener {
-    void onDelete(PostInfo postInfo);
-    void onModify();
-}
